pyproject.toml
[tool.poetry]
name = "fastapi-blog-api"
version = "0.1.0"
description = ""
authors = ["Vlad Zavorotnyy <vlad.zavorotnyy@gmail.com>"]
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
sqlalchemy = "^2.0.29"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
pydantic-settings = "^2.2.1"
python-jose = "^3.3.0"
passlib = "^1.7.4"
celery = "^5.4.0"
kombu = "^5.3.0"
gunicorn = "^21.2.0"
boto3 = "^1.34.69"
python-multipart = "^0.0.9"
jinja2 = "^3.1.3"
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"
pytest-cov = "^4.1.0"
pydantic = { version = "^2.10.6", extras = ["email"] }

[tool.poetry.group.dev.dependencies]
mypy = "^1.4.1"
httpx = "^0.24.1"
pytest-xdist = "^3.6.1"
pytest-mock = "^3.14.0"
pre-commit = "^3.8.0"

[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
ignore_missing_imports = true
strict_optional = true
plugins = ["sqlalchemy.ext.mypy.plugin"]

[tool.ruff]
target-version = "py312"
line-length = 88
lint.ignore = ["D", "ANN", "INP", "FAST", "B008", "PT003", "RET504"]
